backend/requirements.txt
aerich==0.7.1
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
click==8.1.4
colorama==0.4.6
dictdiffer==0.9.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
fastapi-filter==0.6.0
h11==0.14.0
httptools==0.6.0
idna==3.4
iso8601==1.1.0
passlib==1.7.4
psycopg-pool==3.1.7
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pydantic==2.0.2
pydantic-settings==2.0.1
pydantic_core==2.1.2
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomlkit==0.11.8
tortoise-orm==0.19.3
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
